Fix module path
Fixes #5
